fix(roster): Don't leave wrong __opts__ hanging around
Previously, the __opts__ during state rendering were sometimes modified,
similar to the x509_v2 file test mode issue.

This uses a whole new interpreter to avoid such issues. It has the
tradeoff that now, having `salt-run` in the PATH is required.

This can be improved upon, but it's a quick fix for an annoying
problem.